app/requirements.txt
fastapi==0.109.1
uvicorn[standard]==0.24.0.post1
sqlmodel==0.0.16
alembic==1.12.1
pydantic_settings==2.2.1
psycopg==3.1.13
tenacity==8.2.3
python-jose==3.3.0
cryptography==43.0.0
python-multipart==0.0.9